requirements.txt
flask
flask-cors
tensorflow
waitress
pillow